safari-600.7-branch/LayoutTests/http/tests/security/cross-frame-access-protocol-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "https://127.0.0.1:8443".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "https". Protocols must match.


PASS: Cross frame access to https from http was denied!


--------
Frame: 'aFrame'
--------
Inner iframe.
